backend/requirements.txt
django==3.0.4
djangorestframework==3.11.0
django-celery-results== 1.2.1
django-filter==3.2.4
django-celery